requirements.txt
numpy
openai
python-dotenv
flask